requirements.txt
absl-py==2.3.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
aiosqlite==0.21.0
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
astunparse==1.6.3
attrs==25.4.0
banks==2.2.0
beautifulsoup4==4.14.2
blinker==1.9.0
cachetools==6.2.1
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
dataclasses-json==0.6.7
deepface==0.0.95
defusedxml==0.7.1
Deprecated==1.2.18
dirtyjson==1.0.8
distro==1.9.0
fastapi==0.120.4
filelock==3.20.0
filetype==1.2.0
fire==0.7.1
Flask==3.1.2
flask-cors==6.0.1
flatbuffers==25.9.23
frozenlist==1.8.0
fsspec==2025.10.0
gast==0.6.0
gdown==5.2.0
google-ai-generativelanguage==0.9.0
google-api-core==2.28.1
google-api-python-client==2.186.0
google-auth==2.42.1
google-auth-httplib2==0.2.1
google-pasta==0.2.0
googleapis-common-protos==1.71.0
greenlet==3.2.4
griffe==1.14.0
grpcio==1.76.0
grpcio-status==1.71.2
gunicorn==23.0.0
h11==0.16.0
h5py==3.15.1
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.36.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.11.1
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
keras==3.12.0
langchain==1.0.3
langchain-classic==1.0.0
langchain-community==0.4.1
langchain-core==1.0.2
langchain-google-genai==3.0.0
langchain-text-splitters==1.0.0
langgraph==1.0.2
langgraph-checkpoint==3.0.0
langgraph-prebuilt==1.0.2
langgraph-sdk==0.2.9
langsmith==0.4.39
libclang==18.1.1
lz4==4.4.5
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.1
mdurl==0.1.2
ml_dtypes==0.5.3
mpmath==1.3.0
mtcnn==1.0.0
multidict==6.7.0
mypy_extensions==1.1.0
namex==0.1.0
nest-asyncio==1.6.0
networkx==3.5
numpy==2.2.6
opencv-python==4.12.0.88
opt_einsum==3.4.0
optree==0.17.0
orjson==3.11.4
ormsgpack==1.11.0
packaging==25.0
pandas==2.2.3
pillow==10.4.0
platformdirs==4.5.0
propcache==0.4.1
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.3
pydantic-settings==2.11.0
pydantic_core==2.41.4
Pygments==2.19.2
pyparsing==3.2.5
pypdf==6.1.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
regex==2025.10.23
requests==2.32.5
requests-toolbelt==1.0.0
retina-face==0.0.17
rich==14.2.0
rsa==4.9.1
safetensors==0.6.2
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.44
starlette==0.49.3
striprtf==0.0.26
sympy==1.14.0
tenacity==9.1.2
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.2.0
tf_keras==2.20.1
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.1
torch==2.9.0
torchaudio==2.9.0
torchvision==0.24.0
tqdm==4.67.1
transformers==4.57.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.38.0
Werkzeug==3.1.3
wrapt==1.17.3
xxhash==3.6.0
yarl==1.22.0
zstandard==0.25.0
